Updated test suite web test
